Plaster some objects over three.js
The entry point is SphereWorld.World()

Create a Sphere object and use World.add() to add it to the world.

You pass a function to World.draw() that is meant to run
on every frame.

You run the simulation with World.run(), and  World.runDemo() runs
the sphere gravity demo.